backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
pymongo==4.6.0
python-dotenv==1.0.0
